-- fixing error for councils and announcement

+++ app/Http/Controllers/Admin/CouncilsController.php
@@ -53,13 +53,16 @@ class CouncilsController extends Controller
                             'u2.name as reserve_name')
                         ->get();
         else
-            $councils = Council::join('users', 'councils.user_id', '=', 'users.id')
+            $councils = Council::join('users as u1', 'councils.user_id', '=', 'u1.id')
+                        ->join('users as u2', 'councils.reserve_id', '=', 'u2.id')
                         ->where('user_id', '=', Auth::user()->id)
+                        ->orWhere('reserve_id', '=', Auth::user()->id)
                         ->select(
                             'councils.id as id', 'councils.name as name', 'councils.city as city',
                             'councils.account as account', 'councils.maticni as maticni',
                             'councils.latitude as latitude', 'councils.longitude as longitude',
-                            'councils.pib as pib', 'councils.phone as phone', 'users.name as user_name')
+                            'councils.pib as pib', 'councils.phone as phone', 'u1.name as user_name',
+                            'u2.name as reserve_name')
                         ->get();
 
         return view('admin.councils.allcouncils', ['active' => 'allCouncils', 'councils' => $councils]);
@@ -392,8 +395,12 @@ class CouncilsController extends Controller
 
 
         $document = Document::where('type_id', '=' , $announcement->id)->first();
-        File::delete(public_path('documents/'.$document->name));
-        $document->delete();
+
+        if($document != null) {
+            File::delete(public_path('documents/' . $document->name));
+            $document->delete();
+        }
+
         $announcement->delete();
 
         Session::flash('acttab', 'announcements');
